pyproject.toml
[tool.poetry]
name = "py-nova"
version = "0.1.0"
description = "A simple python implementation of Nova: Recursive Zero-Knowledge Arguments from Folding Schemes"
authors = ["PO <wwwcjf163@163.com>"]
readme = "README.md"
packages = [{include = "py_nova"}]

[tool.poetry.dependencies]
python = "^3.11"
py-ecc = "^6.0.0"
pycryptodomex = "^3.17"
numpy = "^1.24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
